requirements.txt
numpy
torch
torchaudio
transformers
librosa
pandas
wav2clip
speechbrain
s3prl
plotly
kaleido
seaborn
scikit-learn==1.5.1
umap-learn